Split out more JSX out of .js files; update ComponentListItem to not access refs in render and update props
