+++ editor/src/views/projects/projectView/ProjectComponents/ProjectComponentsListItem.jsx
@@ -85,6 +85,7 @@ const ProjectComponentsListItem = ({
     <ComponentListItem
       key={component.project_component_id}
       component={component}
+      componentRef={component._ref}
       isExpanded={isExpanded}
       onZoomClick={onZoomClick}
       onListItemClick={onListItemClick}
