+++ editor/src/views/projects/projectsListView/ProjectsListViewTable.jsx
@@ -20,10 +20,10 @@ import { useSearch } from "src/views/projects/projectsListView/useProjectListVie
 import { useAdvancedSearch } from "src/views/projects/projectsListView/useProjectListViewQuery/useAdvancedSearch";
 import ProjectListViewQueryContext from "src/components/QueryContextProvider";
 import {
-  useCsvExport,
-  CsvDownloadingDialog,
   CsvDownloadOptionsDialog,
-} from "src/views/projects/projectsListView/useProjectListViewQuery/useCsvExport";
+  CsvDownloadingDialog,
+} from "src/views/projects/projectsListView/CsvDownloadingDialogs";
+import { useCsvExport } from "src/views/projects/projectsListView/useProjectListViewQuery/useCsvExport";
 import ProjectListToolbar from "src/views/projects/projectsListView/ProjectListToolbar";
 import { useCurrentData } from "src/views/projects/projectsListView/useProjectListViewQuery/useCurrentData";
 import ProjectsListViewMap from "src/views/projects/projectsListView/ProjectsListViewMap";
